Create The Currency Converter By ADNAN & SHAHZEB
This is Adnan Ali & Shahzeb Khan and we have created a program on java. In this program you can easily convert your currency of state in to other currency of state and checkout the worth difference. We have mentioned a few county currencies to be converted in this program like "Rupee, Dollar, Pound, Euro, Yen, and Ringgit.